pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "cli"]

[project]
name = "brahmahub"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "gunicorn>=22.0",
    "asyncpg>=0.30",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "google-genai>=1.0",
    "Pillow>=10.0",
    "psycopg2-binary>=2.9",
    "click>=8.0",
    "rich>=13.0",
    "pyjwt[crypto]>=2.8",
]

[project.scripts]
ihub = "cli.ingesthub_cli.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.14",
    "ty>=0.0.8",
]

# ── pytest ─────────────────────────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# ── ruff ───────────────────────────────────────────────────────

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
extend-select = [
    "I",       # isort
    "B",       # flake8-bugbear
    "UP",      # pyupgrade
    "SIM",     # flake8-simplify
    "RUF",     # ruff-specific
    "ASYNC",   # flake8-async
    "FAST",    # FastAPI
    "PERF",    # Perflint
    "T20",     # flake8-print
    "PTH",     # flake8-use-pathlib
    "PT",      # flake8-pytest-style
    "S",       # flake8-bandit (security)
]
ignore = [
    "FAST002",  # FastAPI Annotated — TODO: migrate endpoints incrementally
    "ASYNC240", # async pathlib — existing code uses sync Path in async handlers
    "T20",      # print() — CLI uses print intentionally
    "B008",     # Query() in defaults — standard FastAPI pattern
    "SIM105",   # contextlib.suppress — style preference
    "SIM108",   # ternary operator — style preference
    "RUF005",   # collection unpacking — style preference
    "PERF401",  # list.extend — style preference
    "S104",    # hardcoded-bind-all-interfaces — 0.0.0.0 is intentional in dev/Docker
    "S608",    # SQL injection — asyncpg uses $N params; f-strings only interpolate column/table names from code
]

[tool.ruff.lint.per-file-ignores]
"cli/**" = ["B905", "S110", "S603", "S607"]  # zip strict, except-pass, subprocess — CLI scripts
"api/routers/system.py" = ["S603", "S607"]       # subprocess — self-update runs git
"api/services/github_auth.py" = ["S310"]          # urlopen — GitHub API calls for token exchange
"tests/**" = ["S101", "S106", "S108"]            # assert, hardcoded values, /tmp paths are expected in tests

# ── ty ─────────────────────────────────────────────────────────

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["api", "cli", "tests"]
